fix crash at the first run with -backupwallet

+++ src/init.cpp
@@ -1219,9 +1219,12 @@ bool AppInit2(boost::thread_group& threadGroup, CScheduler& scheduler)
                 }
 
                 LogPrintf("Backing up wallet %s to %s\n", walletFile, walletFile + ".bak." + currDate);
-                boost::filesystem::copy_file(walletFile,
-                    walletFile + ".bak." + currDate,
-                    boost::filesystem::copy_option::overwrite_if_exists);
+                if(boost::filesystem::exists(walletFile))
+                {
+                    boost::filesystem::copy_file(walletFile,
+                        walletFile + ".bak." + currDate,
+                        boost::filesystem::copy_option::overwrite_if_exists);
+                }
             }
         }
